Move deploy contracts method to chain utils
