+++ node/engine/chainservice/anvil_chain.go
@@ -2,21 +2,15 @@ package chainservice
 
 import (
 	"context"
-	"fmt"
 	"math/big"
 	"os"
 	"os/exec"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	ethTypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
-	NitroAdjudicator "github.com/statechannels/go-nitro/node/engine/chainservice/adjudicator"
-	ConsensusApp "github.com/statechannels/go-nitro/node/engine/chainservice/consensusapp"
 	chainutils "github.com/statechannels/go-nitro/node/engine/chainservice/utils"
-	VirtualPaymentApp "github.com/statechannels/go-nitro/node/engine/chainservice/virtualpaymentapp"
-	"github.com/statechannels/go-nitro/types"
 )
 
 const (
@@ -31,19 +25,13 @@ var ChainPks = []string{
 	"5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a",
 }
 
-type ContractAddresses struct {
-	NaAddress  common.Address
-	VpaAddress common.Address
-	CaAddress  common.Address
-}
-
 type AnvilChain struct {
 	ChainUrl          string
 	ChainAuthToken    string
 	ChainPks          []string
 	AnvilCmd          *exec.Cmd
 	ethClient         *ethclient.Client
-	ContractAddresses ContractAddresses
+	ContractAddresses chainutils.ContractAddresses
 }
 
 func NewAnvilChain() (*AnvilChain, error) {
@@ -62,7 +50,7 @@ func NewAnvilChain() (*AnvilChain, error) {
 		return nil, err
 	}
 	anvilChain.ethClient = ethClient
-	contractAddresses, _ := DeployContracts(context.Background(), ethClient, txSubmitter)
+	contractAddresses, _ := chainutils.DeployContracts(context.Background(), ethClient, txSubmitter)
 	anvilChain.ContractAddresses = contractAddresses
 	return &anvilChain, nil
 }
@@ -92,51 +80,4 @@ func StartAnvil() (AnvilChain, error) {
 	// If Anvil start successfully, delay by 1 second for the chain to initialize
 	time.Sleep(1 * time.Second)
 	return anvilChain, nil
-}
-
-// DeployContracts deploys the NitroAdjudicator, VirtualPaymentApp and ConsensusApp contracts.
-func DeployContracts(ctx context.Context, ethClient *ethclient.Client, txSubmitter *bind.TransactOpts) (ContractAddresses, error) {
-	na, err := deployContract(ctx, "NitroAdjudicator", ethClient, txSubmitter, NitroAdjudicator.DeployNitroAdjudicator)
-	if err != nil {
-		return ContractAddresses{}, err
-	}
-
-	vpa, err := deployContract(ctx, "VirtualPaymentApp", ethClient, txSubmitter, VirtualPaymentApp.DeployVirtualPaymentApp)
-	if err != nil {
-		return ContractAddresses{}, err
-	}
-
-	ca, err := deployContract(ctx, "ConsensusApp", ethClient, txSubmitter, ConsensusApp.DeployConsensusApp)
-	if err != nil {
-		return ContractAddresses{}, err
-	}
-
-	return ContractAddresses{
-		NaAddress:  na,
-		VpaAddress: vpa,
-		CaAddress:  ca,
-	}, nil
-}
-
-type contractBackend interface {
-	NitroAdjudicator.NitroAdjudicator | VirtualPaymentApp.VirtualPaymentApp | ConsensusApp.ConsensusApp
-}
-
-// deployFunc is a function that deploys a contract and returns the contract address, backend, and transaction.
-type deployFunc[T contractBackend] func(auth *bind.TransactOpts, backend bind.ContractBackend) (common.Address, *ethTypes.Transaction, *T, error)
-
-// deployContract deploys a contract and waits for the transaction to be mined.
-func deployContract[T contractBackend](ctx context.Context, name string, ethClient *ethclient.Client, txSubmitter *bind.TransactOpts, deploy deployFunc[T]) (types.Address, error) {
-	a, tx, _, err := deploy(txSubmitter, ethClient)
-	if err != nil {
-		return types.Address{}, err
-	}
-
-	fmt.Printf("Waiting for %s deployment confirmation\n", name)
-	_, err = bind.WaitMined(ctx, ethClient, tx)
-	if err != nil {
-		return types.Address{}, err
-	}
-	fmt.Printf("%s successfully deployed to %s\n", name, a.String())
-	return a, nil
 }
